Add files via upload
Integrated the YouTube API Handler and all relevant dependencies into the master project. Also integrated the YouTube API handler with the video element of the list data structure. Made minor changes to the XML Handler so that XML files could be written successfully from the basic text element of the list data structure.