Add files via upload
Restructured algorithm to make it iterative and provide a ranking result. Implemented a ranked list data structure to store the ranked results from the iterative algorithm. Restructured the way XML files are written to and read from. Enabled the ability to store results within a page in the XML file. Enabled reading/writing all data structure data types to/from an XML file. Altered the way you create an AudioItem so that you no longer need the metadata when creating the item, but instead it finds it for you.